Move Save v1 files
- Move Save v1 files to a `v1` directory so that they can be later deprecated
- Move active save version context to its own file
- Add empty `v2` directory for Save v2
